Fix all sprites vanishing in straight view at max zoom
The sprite size term (camera_zoom << 13) is computed in 32 bits and
wraps once camera_zoom reaches 262144 (the last two zoom stages at
4K resolution) - the negative result makes draw_fastview_mapwho cull
every thing sprite. Widen the intermediate to 64-bit, the same
treatment #5060 gave project_point_helper.

The #5062 depth bias in draw_frontview_thing_on_element computes the
same expression, so it gets the same widening and stays correct at
the zoom levels this fix makes reachable again.

Reproduced on alpha 1.4.0.5252.

+++ src/engine_render.c
@@ -4951,7 +4951,7 @@ static void draw_fastview_mapwho(struct Camera *cam, struct BucketKindJontySprit
         shade_intensity = get_thing_shade(thing);
     shade_intensity >>= 8;
 
-    int size_on_screen = thing->sprite_size * ((camera_zoom << 13) / 0x10000 / pixel_size) / 0x10000;
+    int size_on_screen = thing->sprite_size * (int)((((int64_t)camera_zoom << 13) / 0x10000) / pixel_size) / 0x10000;
     if ( thing->rendering_flags & TRF_Tint_Flags )
     {
         lbDisplay.DrawFlags |= Lb_SPRITE_REMAP;
@@ -8896,7 +8896,7 @@ static void draw_frontview_thing_on_element(struct Thing *thing, struct Map *map
         convert_world_coord_to_front_view_screen_coord(&interp.mappos, cam, &cx, &cy, &cz);
         if (is_free_space_in_poly_pool(1))
         {
-            int size_on_screen = thing->sprite_size * ((camera_zoom << 13) / 0x10000 / pixel_size) / 0x10000;
+            int size_on_screen = thing->sprite_size * (int)((((int64_t)camera_zoom << 13) / 0x10000) / pixel_size) / 0x10000;
             add_thing_sprite_to_polypool(thing, cx, cy, cy, cz - 3 - (size_on_screen >> 1));
             if ((thing->class_id == TCls_Creature) && is_free_space_in_poly_pool(1))
             {
